Updating the Terrain class
The Terrain is a succession of tiles. The tiles are sf::RectangleShape. The terrain can then be defined as an std::vector<int>, where the elements are the height at which a tile of the terrain should be.

For instance, a flat terrain at the height h is represented as {h, h, h, h, h, h, h, h, h}